requirements.txt
Django==5.0.4
Django==5.1.6
fitz==0.0.1.dev2
numpy==2.2.4
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
opencv_python==4.11.0.86
pdf2image==1.17.0
Pillow==11.2.0
pycryptodome==3.22.0
pypdf==5.4.0
reportlab==4.3.1
scikit_learn==1.6.1
spacy==3.8.4
